requirements.txt
dotenv==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-whois==0.9.5
six==1.17.0
